Changed prototype of llapi_hsm_import
The stripe information should all be uint64_t, since that is what the
underlying layout API uses.

+++ include/lustre/lustre.h
@@ -371,8 +371,8 @@ int llapi_hsm_action_get_dfid(const struct hsm_copyaction_private *hcp,
 			      lustre_fid *fid);
 int llapi_hsm_action_get_fd(const struct hsm_copyaction_private *hcp);
 int llapi_hsm_import(const char *dst, int archive, const struct stat *st,
-		     unsigned long long stripe_size, int stripe_offset,
-		     int stripe_count, int stripe_pattern,
+		     uint64_t stripe_size, uint64_t stripe_offset,
+		     uint64_t stripe_count, uint64_t stripe_pattern,
 		     const char *pool_name, lustre_fid *newfid);
 const struct hsm_action_item *llapi_hsm_hai_first(const struct hsm_action_list *hal);
 const struct hsm_action_item *llapi_hsm_hai_next(const struct hsm_action_item *hai);
